Refactor each section in its own component
